+++ superset/exceptions.py
@@ -1,0 +1,30 @@
+# -*- coding: utf-8 -*-
+
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
+
+
+class SupersetException(Exception):
+    status = 500
+
+
+class SupersetTimeoutException(SupersetException):
+    pass
+
+
+class SupersetSecurityException(SupersetException):
+    pass
+
+
+class MetricPermException(SupersetException):
+    pass
+
+
+class NoDataException(SupersetException):
+    status = 400
+
+
+class SupersetTemplateException(SupersetException):
+    pass
