+++ superset/viz.py
@@ -37,6 +37,7 @@ from six.moves import cPickle as pkl, reduce
 from superset import app, cache, get_manifest_file, utils
 from superset.utils import DTTM_ALIAS, merge_extra_filters
 
+
 config = app.config
 stats_logger = config.get('STATS_LOGGER')
